// src/types.ts
export enum SlideLayout {
  WITHOUT = "without",
  TOP_IMAGE = "top-image",
  LEFT_IMAGE = "left-image",
  RIGHT_IMAGE = "right-image",
  TITUL = "titul",
  SECTION_INTRO = "section-intro",
  FINAL = "final",
}

export enum SlideVerticalAlign {
  TOP = "top",
  CENTER = "center",
  BOTTOM = "bottom",
}

export type Slide = {
  id: string;
  title?: string;
  content: string;
  verticalAlign?: SlideVerticalAlign;
  colorAccent?: "light" | "dark";
  backgroundImageUrl?: string;
  layout?: SlideLayout;
  layoutImageUrl?: string;
  isLoadingLayoutImage?: boolean;
  index?: number;
};

export type Presentation = {
  id: string;
  title: string;
  slides: Slide[];
  themeId: string;
  planId?: string | null;
  isDeleted?: boolean;
  isShared?: boolean;
  updatedAt?: Date;
  createdAt?: Date;
};

export enum SlideAction {
  CHANGE_LAYOUT = "change-layout",
  SPELL_CHECK = "spell-check",
  TRANSLATE_TO_RUSSIAN = "translate-to-russian",
  WRITE_MORE_DETAILED = "write-more-detailed",
  SHORTEN_TEXT = "shorten-text",
  SPLIT_INTO_ITEMS = "split-into-items",
  SPLIT_INTO_SECTIONS = "split-into-sections",
  IMPROVE_TEXT = "improve-text",
}

export enum SlidesTemplates {
  EMPTY = "empty",
  TWO_COLS_WITH_SUBHEADINGS = "twoColsWithSubheadings",
  TWO_COLS = "twoCols",
  TWO_COLS_WITH_BAR_CHART = "twoColsWithBarChart",
  THREE_COLS = "threeCols",
  FOUR_COLS = "fourCols",
  CARDS = "cards",
  FRONT_SLIDE = "frontSlide",
  IMAGE_WITH_TEXT = "imageWithText",
  TEXT_WITH_IMAGE = "textWithImage",
  TITLE_WITH_LIST_OPTIONS_AND_IMAGE = "titleWithListOptionsAndImage",
  TITLE_WITH_LIST_OPTIONS = "titleWithListOptions",
  TITLE_WITH_TABLE = "titleWithTable",
  TITLE_WITH_FEATURES_LIST = "titleWithFeaturesList",
  TITLE_WITH_TIMELINE = "titleWithTimeline",
  ARROWS_HORIZONTAL = "arrowsHorizontal",
  PYRAMID = "pyramid",
  STATISTICS = "statistics",
  BIG_NUMBERS = "bigNumbers",
  RATING_STARS = "ratingStars",
  QUOTES = "quotes",
  BENTO_GRID = "bentoGrid",
}